% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{A Gibbs sampler using Rcpp}
\usage{
gibbsC(N, thin, n, a, b)
}
\arguments{
\item{N}{the number of samples}

\item{thin}{the number of between-sample random numbers}

\item{n}{the parameter of the conditional distribution(binomal) of x}

\item{a}{the first parameter of the conditional distribution(beta) of y}

\item{b}{the second parameter of the conditional distribution(beta) of y}
}
\value{
generated samples \code{mat}
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
a <- b <- 5;n <- 20;N <- 1000
gibR <- gibbsR(N, 10, n, a, b)
par(mfrow=c(2,1));
plot(gibR[,1],type='l')
plot(gibR[,2],type='l')
}
}
