% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRS.R
\name{sirs}
\alias{sirs}
\title{SIRS algorithm}
\usage{
sirs(X, y, a = 0.1)
}
\arguments{
\item{X}{the design matrix}

\item{y}{the response}

\item{a}{the parameter selected for penalty rho_a}
}
\value{
xp the estimate of the coefficient \code{xp}
}
\description{
an algorithm for sparse recovery.
}
\examples{
\dontrun{
library(MASS)
set.seed(12345)
n <- 35
p <- 200
beta_0 <- numeric(p)
beta_0[1:7] <- c(1, -0.5, 0.7,-1.2, -0.9, 0.3, 0.55)
r <- 0.1
gamma_r <- matrix(r, p, p) + diag(1 - r, p, p)
X <- mvrnorm(n, numeric(p), gamma_r)
y <- X\%*\%beta_0
beta_hat <- sirscore(X, y, 0.3)
print(sum(beta_hat != 0))
beta_hat <- sirs(X, y, 0.3)
print(sum(beta_hat != 0))
}
}
