\name{StatComp21057-package}
\alias{StatComp21057-package}
\docType{package}
\title{
  An implementation of sirs algorithm.
}
\description{
  The package contains all of my homework for statistical computing and an implementation of sirs algorithm, which devotes to sparse recovery.
}
\author{
Tan Xin
Maintainer: Tan Xin <tanxin@mail.ustc.edu.cn>
}
\references{
  A UNIFIED APPROACH TO MODEL SELECTION AND SPARSE RECOVERY USING REGULARIZED LEAST SQUARES.
}
\keyword{ package }
